Refactor multer
