+++ server/app.js
@@ -2,9 +2,7 @@ const express = require("express");
 const bodyParser = require("body-parser");
 //nanti disini require route
 const beritaRoutes = require('./app/route/berita')
-const multer = require("multer");
 const path = require("path");
-const { v4: uuidv4 } = require('uuid');
 
 const app = express();
 
@@ -17,47 +15,9 @@ const Associations = require("./app/util/associations");
 
  
 
-/*
-@author 14 KP
-Untuk menentukan file gambarnya ingin disimpan dimana 
-*/
-const fileStorage = multer.diskStorage({
-    destination: (req, file, cb) => {
-        cb(null, 'app/public/images');
-    },
-    filename: (req, file, cb) => {
-       const fileExtension = "." + file.mimetype.toString().split("/")[1];
-        cb(null, uuidv4()+ fileExtension)
-    }
-});
-
-/*
-@author 14 KP
-Untuk memfilter tipe file
-*/
-const fileFilter = (req, file, cb) => {
-  if (
-    file.mimetype === "image/png" ||
-    file.mimetype === "image/jpg" ||
-    file.mimetype === "image/jpeg"
-  ) {
-    cb(null, true);
-  } else {
-    cb(null, false);
-  }
-};
 
 app.use(bodyParser.json());
-/*
-@author 14 KP
-membuat middleware dengan storage dan file filter yang telah didefinisikan sebelumnya dan hanya bisa upload satu gambar
-*/
-app.use(
-  multer({
-    storage: fileStorage,
-    fileFilter: fileFilter,
-  }).single("url_gambar")
-);
+
 
 /*
 @author 14 KP
